[refactor] MemberEntity 내부 Payment, Subscription 디렉토리 분리
