+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Member/MemberEntity.java
@@ -7,6 +7,9 @@ import com.example.LifeMaster_BE.Report.ReportEntity;
 import com.example.LifeMaster_BE.SelfDevelop.note.diary.DiaryEntity;
 import com.example.LifeMaster_BE.SelfDevelop.note.thank.ThankEntity;
 import com.example.LifeMaster_BE.TimeManager.Alarm.AlarmEntity;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentEntity;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentStatus;
+import com.example.LifeMaster_BE.UserManager.Member.Subscription.SubscriptionPlan;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
