+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Member/Payment/PaymentDto.java
@@ -1,4 +1,4 @@
-package com.example.LifeMaster_BE.UserManager.Member;
+package com.example.LifeMaster_BE.UserManager.Member.Payment;
 
 import lombok.Getter;
 import lombok.Setter;
