+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Member/Payment/PaymentEntity.java
@@ -1,5 +1,6 @@
-package com.example.LifeMaster_BE.UserManager.Member;
+package com.example.LifeMaster_BE.UserManager.Member.Payment;
 
+import com.example.LifeMaster_BE.UserManager.Member.MemberEntity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Getter;
