+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Member/Payment/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.example.LifeMaster_BE.UserManager.Member;
+package com.example.LifeMaster_BE.UserManager.Member.Payment;
 
 public enum PaymentStatus {
     UNPAID,   // 미결제
