+++ LifeMaster-BE/src/main/java/com/example/LifeMaster_BE/UserManager/Member/Subscription/MemberSubscriptionService.java
@@ -1,5 +1,11 @@
-package com.example.LifeMaster_BE.UserManager.Member;
+package com.example.LifeMaster_BE.UserManager.Member.Subscription;
 
+import com.example.LifeMaster_BE.UserManager.Member.MemberEntity;
+import com.example.LifeMaster_BE.UserManager.Member.MemberRepository;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentDto;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentEntity;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentRepository;
+import com.example.LifeMaster_BE.UserManager.Member.Payment.PaymentStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
